requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pandas
python-dotenv
requests
matplotlib
jinja2
